Update optimizer.js
Updated the optimizer.js
Now it works for the proposed error scenario.
I believe it is not precisely the thing that was expected, but it works. I just do not understand how elegantly use .keys() and .values() to manipulate and than structure back in two-layer array to construct the object back with from.Entries().